feat(crawler): Unpaywall APIを利用してPDF URLを取得する機能を追加 (#25)
* feat(crawler): Unpaywall APIを利用してPDF URLを取得する機能を追加

- usecase/unpaywall.py: Unpaywall APIを使用して論文のPDF URLを取得するUnpaywallSearchクラスを実装
- main.py: Semantic Scholarによる処理の後、Unpaywallを使用してPDF URLを補完する処理を追加
- configs: APIリクエストに必要なメールアドレス設定を追加
- tests/test_unpaywall.py: UnpaywallSearchに対するユニットテストを追加
- usecase/semantic_scholar.py: APIパスの定義を定数化し、データ取得ロジックをリファクタリング

* refactor(crawler): update enrich_papers to modify objects in-place and add overwrite option

- Update `SemanticScholarSearch.enrich_papers` and `UnpaywallSearch.enrich_papers` to modify the input list of `Paper` objects in-place instead of returning a new list.
- Ensure all input papers are returned, maintaining the original order and count, even if API data is missing.
- Add `overwrite` parameter to `enrich_papers` methods to control whether existing metadata (Abstract, PDF URL) should be overwritten by API results.
- Improve PDF URL resolution logic in Unpaywall to fallback to `oa_locations` if `best_oa_location` is missing.
- Add `LOG_LEVEL` configuration and apply it to the logger setup.
- Fix typos in `DBLPSearch` error message and `UnpaywallSearch` path constant.
- Update type hint for `Crawl-delay` in `RobotGuard`.
- Update tests to reflect in-place modifications and cover new `overwrite` behavior.

* fix(crawler): handle API errors gracefully in Semantic Scholar and Unpaywall

- Catch httpx.HTTPStatusError in Semantic Scholar and Unpaywall fetch methods to prevent the entire process from crashing on a single failure.
- Log 404 errors as debug and other HTTP errors as warnings, then return None instead of raising an exception.
- Update enrichment logic to ignore None results from failed fetch attempts.
- Modify unit tests to assert that API errors result in unenriched paper objects rather than exceptions.

* Update workflows/crawler/usecase/semantic_scholar.py


* Update workflows/crawler/libs/log.py


---------

+++ workflows/crawler/libs/__init__.py
@@ -102,7 +102,7 @@ class RobotGuard:
         設定がない場合はNoneを返します。
 
         Returns:
-            float | None: Crawl-delay(秒数)。設定がない場合はNone
+            int | None: Crawl-delay(秒数)。設定がない場合はNone
 
         Raises:
             RuntimeError: robots.txtがまだロードされていない場合
